requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
requests
beautifulsoup4
fastapi
pydantic
